novo GATE para verificação, e tradução do projeto

+++ app/Http/Controllers/FuncionarioController.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreUpdateFuncionarios;
 use App\Models\Funcionario;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Input;
+
 
 class FuncionarioController extends Controller
 {
@@ -33,13 +36,29 @@ class FuncionarioController extends Controller
     }
 
     public function store(Funcionario $funcionario, StoreUpdateFuncionarios $request) {
-       
+        $rules = [
+            'name' => 'required|min:3|max:100',
+            'status' => 'required',
+            'section' => 'required|min:3|max:100',
+            'function' => 'required|min:3|max:100'
+        ];
+
+        $validator = Validator::make($request->all(), $rules);
+
+        if($validator->fails()) {
+            return redirect('funcionario.create')
+                   ->withErrors($validator)
+                   ->withInput();
+        }
+
         $data = $request->all();
         $data['user_id'] = Auth::user()->id;
         
         $funcionario = Funcionario::create($data);
 
-        return redirect()->route('funcionario.index')->with('message', 'Funcionário ADICIONADO com sucesso');
+        return redirect() 
+               ->route('funcionario.index')
+               ->with('message', 'Funcionário ADICIONADO com sucesso');
     }
     
     public function show($id) {
@@ -54,32 +73,39 @@ class FuncionarioController extends Controller
 
     public function edit($id, Funcionario $funcionario) {
         $funcionario = Funcionario::find($id);
-
-        if(!$funcionario)
-        return redirect()->back()->withErrors($validator);
-
-        return view('funcionario.edit', compact('funcionario'));
+        $user = Auth::user()->id;
+        if( $user === $funcionario->user_id) {
+            if(!$funcionario){
+                return redirect()->back()->withErrors($validator);
+            }
+            return view('funcionario.edit', compact('funcionario'));
+        } else {
+            return redirect()->route('funcionario.index');
+        }
     }
 
     public function update (Funcionario $funcionario, StoreUpdateFuncionarios $request, $id) {
-        if(!$funcionario = Funcionario::find($id)) {
-            return redirect()->back()->withErrors($validator)->withInput();  
-        } else {
-            $data = $request->all();
-            $data['user_id'] = Auth::user()->id;
-            $funcionario->update($data);
-
-            return redirect()->route('funcionario.index')->with('message', 'Funcionário ALTERADO com sucesso');
-        }
         
+            if(!$funcionario = Funcionario::find($id)) {
+                return redirect()->back()->withErrors($validator)->withInput();  
+            } else {
+                $data = $request->all();
+                $data['user_id'] = Auth::user()->id;
+                $funcionario->update($data);
+    
+                return redirect()->route('funcionario.index')->with('message', 'Funcionário ALTERADO com sucesso');
+            }
     }
 
     public function destroy($id, Funcionario $funcionario) {
         $funcionario = Funcionario::findOrFail($id);
-
-        $funcionario->delete();
-
-        return redirect()->route('funcionario.index')->with('message', 'Funcionário DELETADO com sucesso');
+        $user = Auth::user()->id;
+        if($user === $funcionario->user_id) {
+            $funcionario->delete();
+            return redirect()->route('funcionario.index')->with('message', 'Funcionário DELETADO com sucesso');
+        }else {
+            return redirect()->route('funcionario.index');
+        }
     }
 
 }
